refactor: 💡 Changing the directory and import destination
